requirements.txt
click>=8.1.0
librosa>=0.10.0
soundfile>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pydub>=0.25.1
numpy>=1.24.0
scipy>=1.11.0

